+++ LargeInt/test/comparison.test.cpp
@@ -6,10 +6,10 @@
 #include "LargeIntImplementation.h"
 
 TEST(large_int, comparison) {
-    LargeNumbers::LargeIntImplementation<uint8_t> integer("0");
-    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint8_t>("0")), 0);
-    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint8_t>("-1")), 1);
-    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint8_t>("1")), -1);
-    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint8_t>("256")), -1);
-    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint8_t>("-256")), 1);
+    LargeNumbers::LargeIntImplementation<uint32_t> integer("0");
+    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint32_t>("0")), 0);
+    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint32_t>("-1")), 1);
+    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint32_t>("1")), -1);
+    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint32_t>("256")), -1);
+    EXPECT_EQ(integer.compare(LargeNumbers::LargeIntImplementation<uint32_t>("-256")), 1);
 }
